+++ CHANGES.md
@@ -6,6 +6,7 @@ develop
 - implement assertTrueAndSay() and assertFalseAndSay() with optional message
 - assertEqualsAndSay(...) takes Object as parameters and has optional message
 - removed "important" from assert...AndSay() output
+- create Response object to avoid consumers of this library having a dependency on the ApiTester library.
 
 Version 0.2.0: 09.07.2012
 -------------------------